Migrate reference/roles to v2, change updated field format to iso8601 (#25)
Fixes an inconsistent datetime format bug in reference/roles. Left the
existing endpoint untouched, copied most of its functionality to
/v2/reference/roles and made the datetime formats consistent.

Testing:
For volunteers and supervisors, GET request should return 
```
[
    {
        "id": 1,
        "code": "basic",
        "name": "Basic",
        "created": "2021-04-10T10:03:32",
        "updated": "2021-04-10T10:03:32",
        "deleted": "False"
    },
    {
        "id": 2,
        "code": "driver",
        "name": "Driver",
        "created": "2021-04-10T10:03:32",
        "updated": "2021-04-10T10:03:32",
        "deleted": "False"
    },
    {
        "id": 3,
        "code": "crewLeader",
        "name": "Crew Leader",
        "created": "2021-04-10T10:03:32",
        "updated": "2021-04-10T10:03:32",
        "deleted": "False"
    },
    {
        "id": 4,
        "code": "advanced",
        "name": "Advanced",
        "created": "2021-04-10T10:03:32",
        "updated": "2021-04-10T10:03:32",
        "deleted": "False"
    }
]
```

For supervisors only, the POST method should return the id of the new
role in the database, and the new role should be visible in the `roles`
table. If the request is malformed then the method returns a `null` id
and does not modify the database.
Example post body:
```
{
    "name": "Test Role",
    "code": "testRole"
}
```

Example response:
```
{
    "id": "8"
}
```
Volunteers are not allowed to use the patch, post or delete methods.

Unauthenticated users should not be able to access any of the methods.

---------

Co-authored-by: Rachel Schroder <u6301105@anu.edu.au>
Co-authored-by: RitaZ <126477233+Xinyun-Zhou@users.noreply.github.com>