Fix bugs in dietary requirements refactor (#29)
Pull Request #27 introduced broken import to main branch. Removed import
because it is not necessary.

Authentication decoding did not work with how Postman creates the bearer
token header, added logic to more robustly decode the token.

Dietary requirements POST method was returning dietary requirements
object with null fields. Frontend does not need this object returned, so
the method now returns only a success field with true or false and a
200/400 code.

Remove user_id as a required field in the POST request body.

---------

Co-authored-by: Rachel Schroder <u6301105@anu.edu.au>