Create testing environment and add Github actions workflow (#192)
## Describe your changes
I've made several changes to the code so that unit tests and functional
tests run in an in-memory SQLite database instead of directly changing
the MySQL application database. Additionally, a GitHub Actions workflow
has been added, but I'm still unsure if it's working correctly.

Note: I made several changes to the legacy code to fix bugs and ensure
compatibility with the SQLite code as well. I don't think it will break
anything in the future.
## Issue ticket number and link

---------

Co-authored-by: Muhammad Hafizh Hasyim <u7316882@anu.edu.au>
Co-authored-by: Yiran Li <u6920084@anu.edu.au>