Edit Unavailability Endpoint (#206)
## Describe your changes
The logic for updating unavailability is simple. Updating any field in
the unavailability with the matching fields in the payload. The most
important part is the payload validation, the payload data is checked in
the repository function and raise custom errors to catch later in the
controller. Some features:
- a user can only edit their own unavailability (this feature will be
later discussed within the meeting to see if it should be applied for
other API endpoints)
- custom errors (EventNotFoundError, InvalidArgumentError)
## Issue ticket number and link